requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiosqlite==0.19.0
python-multipart==0.0.6
jinja2==3.1.2
pyjwt==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
